requirements_pi.txt
flask>=3.0.0
flask-socketio>=5.3.6
paho-mqtt>=1.6.1
opencv-contrib-python>=4.9.0.80
psutil>=5.9.8
requests>=2.31.0
numpy>=1.26.0
RPLCD>=1.3.0
python-socketio[client]>=5.8.0
lgpio>=0.2.2.0
rpi-lgpio>=0.6
